use assertThat in UtilsTest, move to same package as src, remove and inline Utils.doesHaveSeparator method. closes #25
